Refactor into own test file
