+++ test/test_helper.rb
@@ -46,5 +46,34 @@ module Minitest
     def assert_includes_offense(offenses, output)
       assert_includes(offenses.sort_by(&:location).join("\n"), output.chomp)
     end
+
+    class TracerCheck < ThemeCheck::Check
+      attr_reader :calls
+
+      def initialize
+        @calls = []
+      end
+
+      def respond_to?(method)
+        method.start_with?("on_") || method.start_with?("after_") || super
+      end
+
+      def method_missing(method, node)
+        @calls << method
+        @calls << node.value if node.literal?
+      end
+
+      def respond_to_missing?(_method_name, _include_private = false)
+        true
+      end
+
+      def on_node(node)
+        # Ignore, too noisy
+      end
+
+      def after_node(node)
+        # Ignore, too noisy
+      end
+    end
   end
 end
